document-service/requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
pydantic==2.12.3
pydantic_core==2.41.4

# httpx==0.28.1
requests==2.31.0
beautifulsoup4==4.12.3
# playwright==1.40.0
# Database
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.13.1

# Elasticsearch
# elasticsearch==8.11.0


# Environment variables
python-dotenv==1.0.0

# Google Gemini LLM
# google-generativeai==0.3.2


# Caching
redis[async]==5.1.1


annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
h11==0.16.0
idna==3.11
sniffio==1.3.1
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0